Move all high-lvl tests to a single test module
Makes it easier to detect mixed code between high and low-lvl funcs
